+++ scripts/test/Selenium/Agent/Admin/AdminLog.t
@@ -107,6 +107,41 @@ $Selenium->RunTest(
             "Sidebar column is not visible on the screen",
         );
 
+        # Verify log time stamp is in user default time zone (UTC).
+        $Self->True(
+            $Selenium->execute_script("return \$('#LogEntries .Error:eq(0)').text().indexOf('UTC') !== -1"),
+            "Log time stamp is in user default time zone (UTC) format."
+        );
+
+        # Get test UserID.
+        my $UserObject = $Kernel::OM->Get('Kernel::System::User');
+        my $UserID     = $UserObject->UserLookup(
+            UserLogin => $TestUserLogin,
+        );
+
+        # Set test user's time zone.
+        my $UserTimeZone = 'Europe/Berlin';
+        $UserObject->SetPreferences(
+            Key    => 'UserTimeZone',
+            Value  => $UserTimeZone,
+            UserID => $UserID,
+        );
+
+        # Relog test user.
+        $Selenium->Login(
+            Type     => 'Agent',
+            User     => $TestUserLogin,
+            Password => $TestUserLogin,
+        );
+
+        $Selenium->VerifiedGet("${ScriptAlias}index.pl?Action=AdminLog");
+
+        # Verify log time stamp is in user preference time zone (Europe/Berlin).
+        $Self->True(
+            $Selenium->execute_script("return \$('#LogEntries .Error:eq(0)').text().indexOf('$UserTimeZone') !== -1"),
+            "Log time stamp is in user preference time zone ($UserTimeZone) format."
+        );
+
         # Clear log.
         $LogObject->CleanUp();